backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rag-backend"
version = "0.1.0"
description = "RAG SaaS Backend API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "celery[redis]>=5.3.6",
    "redis>=5.0.0",
    "flower>=2.0.1",
    "sqlalchemy[asyncio]>=2.0.30",
    "alembic>=1.13.1",
    "asyncpg>=0.29.0",
    "qdrant-client>=1.9.0",
    "openai>=1.30.0",
    "anthropic>=0.28.0",
    "pypdf>=4.2.0",
    "pdfplumber>=0.11.0",
    "python-docx>=1.1.2",
    "mammoth>=1.7.1",
    "unstructured>=0.14.0",
    "sentence-transformers>=3.0.0",
    "structlog>=24.1.0",
    "pydantic-settings>=2.2.1",
    "python-multipart>=0.0.9",
    "boto3>=1.34.0",
    "tiktoken>=0.7.0",
    "httpx>=0.27.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-httpx>=0.30.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app", "workers"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["app", "workers"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
